Do not generate ObjectIDs from php
Instead fill the _id field with a database generated _id on insertion

I cant find the discussion at this time, but it was discussed that
collisions could happen when creating new instances from local code.

This ensures that the database is responsible for id generation instead
of the client machine.